move testscript at toplevel, avoids stutter. declare we need go1.20 otherwise coverage etc doesn't work
