+++ envvarname_windows.go
