+++ testdata/command.txt
