+++ testdata/commandstatus.txt
