+++ testdata/cpstdout.txt
