+++ testdata/evalsymlink.txt
