+++ testdata/regexpquote.txt
