+++ testdata/stdin.txt
