+++ testdata/testscript_duplicate_name.txt
