+++ testdata/testscript_logging.txt
